✨(core) add user-related API endpoints in core app
We need a `whoami` API endpoint to manage login related tasks on
the frontend, starting with knowing if the current user is logged
in and who they are.

We're using a DRF ViewSet to stay consistent with all our other APIs
but are explicitly not implementing the expected `retrieve` and
`list` endpoints as we have no use for them and they could leak
information.