✨(frontend) add a React-based login/signup area in the header
We need a clean way to check if users are logged in and display
either a signup/login button pair, or the user's name and a log
out button. This without affecting our ability to cache any CMS
page.

We decided to use a React component to solve this issue. With our
new /whoami endpoint, this component can easily check the user's
login status as it bootstraps and display the expected UI.