✨(sandbox) enable auth using the django admin in the sandbox
We want to have basic working links for the login/logout/signup
buttons. Login can be achieved by just directing users to the django
admin. Same thing for signup: users can't really sign up this way but
it's not the sandbox's purpose to enable them.

For logout, we can just add the django contrib auth urls and use the
logout url it gives us.