🐛(api) fix the retrieve category API endpoint
As we updated our tests from manually calling the ViewSets to
using the test client, we noticed the API endpoint to retrive a
category was broken: it receives a kind parameter, which our
method does not expect.

It's not necessary to respond, but it throws errors if we don't
expect it, we can just add it to our function signature.