🐛(api) handle unexpected calls on categories catchall routes
The categories ViewSet sits on a catchall url so it can receive
requests for any “kind” of category in an intuitive way. This has
a drawback though: unexpected requests might end up in our categories
routes.

We simply want to make sure those are handled properly and return
eg. 404 errors, and do not raise 500s or return any unwanted data.