📝(website) document the frontend build for dependents
Dependents can run their own build for CSS & JS to theme, modify
or extend Richie. This was mostly not documented, be it in the
repository or in the NPM package that we publish to facilitate it.

We wrote some docs to help users figure out own to run it. They also
explain the settings for the JS build.