♿️(footer) fix logo link in footer
The logo link in the footer used an empty alt text on the image, which
is the only content for the link. This is incorrect as there is no
text at all on the link for accessibility tech users.

We can just add `{{ SITE.name }}` like on the logo in the header.