♿️(templates) fix language menu accessibility
Link titles, which are necessary for language switching links to provide
context on what those links actually do, were broken due to an incorrect
use of `{% blocktrans %}`.

We used blocktrans variables to fix them, and took this opportunity to
improve the link titles by adding a different one for the currently
selected language.