👷(website) run prettier on the docs website directory
We were not running prettier on the JS and JSON files in the
website directory, and had no CI job to do that. This can cause
ambiguity in formatting, which we resolved in the codebase by
just running prettier.

Add a CI job to lint the website dir with prettier.