✨(search) add toolbar menu item to bootstrap elasticsearch index
There might be some cases in which a staff user needs to regenerate
the search index from scratch. We propose to trigger it via an Ajax
menu item in the toolbar. Only staff users who were granted the
specific permission to manage the Elasticsearch index can do it.

We do it synchronously for the moment because:
- it should take less than 30 seconds,
- only authorized users can do it and it's rare,
- we want to avoid introducing a dependency to Celery or other
  asynchronous tasks handlers.