💄(search) rework the search view layout
The Search component was placing the title and search field inside
a `search__results` div and pushing the filters pane to the very top
of the page.

We now have a dedicated `search__header` area that helps us make the
layout cleaner and lighter while keeping everything as easy to
access as before.