♻️(frontend) remove act warnings in <Search />
The <Search /> component was producing act-related warnings because
some of our tests were making http requests and never explicitly
waiting for them, resulting in the responses being handled (and state
modified) at an undefined time.

Simply adding `await wait()` after rendering makes the test explicitly
wait for resolution and wrap that in an `act()` call (that's what RTL's
`wait()` does), and gets rid of the warnings.