✨(search) rework filter definitions to add a static defs method
We're going to provide a new API endpoint that lets the frontend
get configuration related to search and available filters.

The first thing to do is extract the shaping of this static filter
definition to its own method, which we can expose and then reuse
in the faceted definitions method.

We also added base_path to all static definitions to ensure all
our filter definitions have the same shape.