✨(search) add an API route to get the filter definitions
Some components in the frontend need simple, static versions of the
filter definitions to configure themselves. They are using the main
course search right now but this creates awkward dependencies.

We're adding a new endpoint to just get this static versions of
filter definitions. This will help us clean up the dependencies and
decouple our frontend components at the cost of an additional but
lighter API request.