🏷️(frontend) separate static & faceted filter definition types
Until now, we only had one type of filter definition. We now have
two: the static ones that can be requested on the dedicated endpoint
to bootstrap our components, and the faceted ones that are part of
search results.

We can simply split our filter into a FilterDefinition with just the
static fields and a FacetedFilterDefinition with contextual info based
on the current search.