🏷️(filters) update filter definition type to account for position
Filter definitions (including our new, static versions) include a
"position" property that should be used to order filters in the
filters pane. Until now, it was missing in our types and we were
failing to use it.

The first step to fixing this is to complete our types and our
corresponding examples in tests.