♻️(frontend) refactor search fields with useStaticFilters
We want to stop making full course search requests just to get
the filter definitions in <RootSearchSuggestField />. We can use
our new hook to just get the static definitions instead.

We also need to start unplugging <SearchSuggestField /> from the
rest of <Search />, of which the first part is to stop using course
search results to configure it and rely on hook there instead.