✨(frontend) isolate history into a separate context and hook
We want to use URLs (by means of query string contents) to synchronize
separate React components together. Unfortunately, calls to
`history.pushState`, which we use to modify the URL as the user
updates their search, do not trigger any events in the browser.

To be able to react and re-render any components that need to be
up-to-date on history changes which do not reload the page, we decided
to handle the history in its own hook-and-context that can live on
top of our component hierarchy and let us share state cleanly.

Other hooks such as our courseSearchParams helper can then use this
history context to keep themselves updated.