♻️(frontend) use React Portals to render separate components
We used to call ReactDOM.render for each and every component wee needed
to render on a page, targeting its own `.richie-react` container. This
was sufficient as long as we did not expect to have several components
on some pages, and did not neeed to share anything betweeen them.

Now that many pages are expected to have two or more components, it
makes a lot more sense to have them live under one React tree. It will
allow us to mutualize some elements like Providers which are shared
for all the components.