♻️(i18n) transform react_locale filter into rfc_5646_locale
We're changing the way the backend communicates localization
information with the frontend. Instead of a specialized data-attr
on each would-be React component, we'll be using the builtin
lang attribute on <html> to pick up the language/locale in the
frontend.

However, the <html> lang element uses BCP47/RFC5646 and not
ISO15897 to specify locales, unlike our React internationalization
library. We're starting by switching our template tag to generate
BCP47/RFC5646 and let the frontend build the string it needs from
that.