♻️(core) use <html> to communitate l10n information to the frontend
We previously used <html lang> to specify the language of the page
as is necessary for accessibility tooling, and separate data-locale
attributes to pass ISO15897 locales to our React frontend.

We decided to instead use a BCP47/RFC4656 locale on <html lang> and
then reuse it on the React side to specify locales. This gives us a
single source of truth when it comes to localization, and relies on
web standards instead of home-made props.