🚚(frontend) rename frontend l10n files fo BCP47/RFC5646 format
Our localization files for the frontend were named in the ISO15897
format. Web standards instead push us to use BCP47/RFC5646 because
it is the supported format for native HTML elements.

The only practical difference for lang + region locales is to use
dashes and not underscores. We also modify our crowdin config so
newer files will be generated with the newly expected names.