🐛(courses) add migrate dependency to update section b4 creating footer
The data migration that moves any existing footer to the new static
placeholder was running before the schema migration on the section plugin
allowed null titles. This was resulting in IntegrityErrors. We did not
see this error in development because the database migrations were run
separately. We did not see it in the CI either because migrations run on
an empty database.