🌐(frontend) do not report errors related to en-US l10n files
Richie's frontend operates on the assumption that en-US is the
default language and is contained in the default messaged as opposed
to localization files that are lazily loaded for other languages.

We don't want to add a special case that does not import() an l10n
file for en-US because we don't want to prevent Richie users from
creating their own and loading it.

Therefore, we decided to load it if it's there, but not report an error
that is in the default and most common case, just the app working as
intended.