♻️(frontend) rework course search params hook to use HistoryContext
Until now, the course search params hook was the authoritative source
for search params in the <Search /> component and therefore the whole
search view as far as the frontend is concerned. This worked well when
we could assume there would not be independent components interacting
with user search.

Now that we need to split <Search /> and have independent components
for user search, we need to defer this centralization to a more generic
provider that can br wrapped around the whole app through <Root />: the
new History hook & context.

We then need to rework useCourseSearchParams from a centralized hook
managing its own state to a simple helper that provides utilities to
interact with history and keep all components updated on history
changes without complexifying their API.