♻️(frontend) update CourseSearchParamsContext consumers
Consumers of the CourseSearchParamsContext now need to directly
call the useCourseSearchParams hook. This will allow them to still
keep the exact same API but simply make use of the hook as a utility
to interact and keep up with search params through the more generic
HistoryContext.

We also had to change our testing strategy from proving a mocked,
test-specific CourseSearchParamsContext, to going up one level with
the HistoryContext. This has the added benefit of integrating our
tests closer with the whole stack thee app is using.