✨(frontend) let courseSearchParams dispatch several actions at a time
Our new architecture for course search requires pushState to happen
in the dispatch call for courseSearchParams, instead of delegating it
to a useEffect.

This causes an issue when a component needs to dispatch more than one
action at a time: we then create as many history entries. We can solve
this by accepting more than one action in our custom dispatch.