♻️(search) extract <SearchSuggestField /> from <Search />
We're extracting <SearchSuggestField /> from <Search /> in order to
give more freedom to Richie users. The purpose here is to easily let
them lay the Search template out as they want, with less constraints.

The trade-off we're making is the CSS is "leaky" as parts of <Search />
actually depend on the container type in the search template. Visual
layout also does not exactly map to what's in the DOM.