✨(richie) bootstrap a tests_e2e folder for Cypress-powered tests
We want to add automated testing for accessibility through axe-core. As
Richie is a hybrid app, we can't just test our React components in Jest
tests. We have to run a Django server and use the pages as a user would.

To that end, we decided to use Cypress, which is a solid and productive
end-to-end testing solution. It is easy to integrate it with axe-core
to check a11y violations.