📦(frontend) update webpack config to build for outdated browsers
Simply using the browserslist in our babel config to target outdated
browsers is not enough to have a bundle that works in IE11. We must
also make sure to transpile the dependencies that need it.

We also had to add a babel plugin specifically to support `query-string`
which does not include any affordance for browser environments and
breaks us by having both modern syntax and commonjs exports instead
of ES2015+ exports.