🏁(frontend) make base.html javascript IE11-proof
We have some JS in base.html for the menu that is visible in CMS pages
(and therefore not powered by react). As this bit of JS does not go
through any processing, we need to directly write it to be IE11-proof
if we want it to run there.