✨(blogposts) display all categories related to blogposts on news page
We want to be able to navigate to any category that is used to tag one
or more blog posts, in one click from the news page. We introduce a
template tag that can work on any type of page related to categories
via a CategoryPlugin in any of its placeholders.