✨(demo) improve News page on demo site
In a recent commit, we have added links to all categories targeted by
one or more blog posts on top of the news page. This was rendering
too many categories on the demo site because we were tagging blog posts
with any of the many subjects. We modify the demo to tag blog posts we
a limited set of categories dedicated to news.