✨(search) add more category data on courses in the course index
In Richie's search engine, users sometimes need to get more than
one icon and to differentiate between them based on their meta
categories, their parents, or the placeholders they came from.

The first step is to index more category-related data on courses.
To do this, we relied on the existing `get_categories` method we
just extended with the placeholder slot that links the category
with the course.