🔥(search) remove "icons" from courses, superseded by "categories_data"
Now that we have all the data required to display categories on course
glimpses in any way available as "categories_data", we do not need a
hardcoded field with assumptions on how one of these categories is
going to be used.

As we moved that logic to the frontend (the actual course glimpse),
we can stop handling it in out indexer.