✨(courses) course.get_categories adds placeholder info on categories
We need to change our ElasticSearch index to add more information
about related categories on courses, which will then be passed on
to our API. Notably, we neeed to know how a given category is linked
to the course (eg. through which placeholder and at which position).

We can achieve this through the course.get_categories() method. Using
annotate from Django's ORM, we can add a `placeholder_slot` field on
categories in the resulting QuerySet.

We noted that this prevents categories from being deduplicated
(looking at it through their pk) if a given category is present in two
different placeholders on the same course. Both instances of the
category we get are now different by their `placeholder_slot` and
`placeholder_position` fields. We made this explicit by adding a test
that shows this edge case.