✨(api) return additional category data on courses
Now that we added relevant and contextual information on linked
categories on course objects in our index, we need to update our
indexer to include the relevant part of that data on courses in
API response (incl. localization).

We also needed to update tests to reflect the new shape of our
indexer courses source and the new expected shape of courses as
returned by the API.