🐛(search) fix an issue with get_picture_info when there is no image
Our simple plugin's `get_picture_info()` helper operated on the
assumption that an Picture has a picture attribute, with a
FileImage.

Depending on how actual users use the back-office, it is possible to
create Picture instances that are missing this. This caused the
helper to raise an error and crash the ES index generation.

We added a test to reproduce this issue and solved it by simply
bailing out at the top if there is no image as `get_picture_info()`
has pretty much nothing to return then.