✅(search) add tests for cases of image-less pictures in courses
We recently updated `get_picture_info()` to not break when it gets
an image-less Picture instance, and instead return None. We decided
to expand tests on methods around courses in the indexer to make
sure everything works properly when there is no icon or no cover
image on a course.