✨(search) get_best_field_language does not raise on NoneType
We're adding new multilingual fields to our indexed courses. Some
of these can be None depending on the course, its categories and
their place in the categories tree.

To keep things as simple as possible, we're making
`get_best_field_language()` more resilient: it now gracefully handles
None values (for the field) and just returns None in this case,
instead of raising a TypeError.