✨(search) update the frontend to use categories_data instead of icon
We added a `"categories_data"` field on courses which allows us to
achieve what we previously did with the icon field, although in a
much more flexible way that should help users build their own course
glimpses without re-coding indexing logic.

Now that this new field is available, we can use it instead of `"icon"`
in the frontend. This will allow us to remove `"icon" soon thereafter.`