♻️(settings) make release, env & sentry_dsn available on settings
Information about release, environment and the sentry_dsn were
already used in Richie, although only in a settings `post_setup`
method. They were not available for any other part of the code
to use.

Refactor them into proper settings to make it easier to use from
anywhere and not just in settings hooks. Here, for example, we'll
need to inject them in a template to pass them to the frontend's
Sentry library.