✨(core) add a context processor for frontend context
Frontend components throughout Richie have common expectations
about the props they need to bootstrap. In addition to their own
props, they all support and/or need additional context.

As we're adding environment, release, and sentry_dsn to this shared
context, we can't handcraft it in each django template that uses
a frontend component anymore.

This is why we're adding a context_processor to easily make all
this data available to the frontend whenever it is used.