🌐(courses) localize fields on the Licence model
The `name` and `content` fields on the Licence model contain user-facing
text. We need both of these to be translatable in the back-office and
localized for end users.

The first step is to introduce a `Django-parler` translation model for
Licence. We're then deprecating the existing, untranslated fields and
replacing them with `TranslatedFields` that interact with the new
translation model through parler.