📦(frontend) add a global RICHIE_VERSION in the frontend bundle
We want to be able to get the version in the frontend from the
package.json instead of the backend context and version. This will
enable us to have the correct version for cases where a dependent
built its own frontend to use overrides.

The first use case for this is to have correct version numbers in
Sentry reports for the frontend even in this case.