♻️(courses) use f-strings in course model queries
We're using string interpolation to build queries for related objects
in course model methods, as those queries can be quite unwieldy with
the model nesting that happens in Django CMS.

f-strings are a much cleaner way to write this than `.format` calls.
As we wrote some queries in this style, we can refactor the other
ones to use it too.