✨(course) add a method to get related categories data on course
In some circumstances in Richie, most notably when we're rendering
a course glimpse, we need additional information on categories
beyond just the list of categories linked to the course.

We also want this information to be easy to use from a template or
to return on the API for direct consumption by the frontend.
Therefore, we added a method on the course model that returns
this additional data in an organized but ready-to-consume format.