♻️(search) improve ES index management command and tests
The tests for the `bootstrap_elasticsearch` management command, and
in particular the `regenerate_indexes` tests, were using a lot of
mocks such that they provided very little confidence in the code
actually working as intended.

We replaced two heavily mocked tests with an entirely non-mocked
one that should cover the same code surface with more confidence.
We also touched up on the code in the index_manager module, eg.
fixing a small bug we encountered and replacing format calls with
f strings.