🔧(search) variabilize ElasticSearch index names through settings
ElasticSearch index names are hardcoded in their respective indexers.
There is a `richie_` prefix that should prevent conflicts with other
applications running on the same ElasticSearch instance, but that
still prevents more than one instance of Richie to run on the same
ElasticSearch, as indices would share names.

We can make this changeable through the settings and an environment
variable, using `richie_`as a default so it's not a breaking change
for existing users.