🐛(search) fix a bug that cause `regenerate_indexes` to crash
In some instances, Richie can end up with an ElasticSearch that
contains broken indices, with each index's expected alias as an
index name. This can happen for example if the signals that are
tasked with updating courses, organizations, etc. are happened
in a state where no bootstrap, no `regenerate_indexes` has happened.

ES then implicitly creates the index for the object we're trying to
update.

This caused `regenerate_indexes` to fail when attempting to create
an alias with the same name as an existing index. Since the existing
index as a problematic name, is broken by nature (it was created
with no content, no mappings, no scripts etc.) and an general indices
regeneration was triggered, we can just delete the broken index and
move on.