⚡️(courses) limit the number of related courses on category pages
As contents scale in a Richie instance, pages that automatically load
all items related to some item tend to become very slow: they may
need to query hundreds of related objects.

This is not a good experience. Especially since users are probably not
interested in a hundred-items-long list of courses on eg. a category
page for example.

Use our settings and new `first_n` filter to limit the number of
courses we show on category pages and show a link to see more using
the Search view.