✨(courses) add a context processor for Richie pages' settings
Richie object detail pages (eg. course pages, category pages, etc.)
need some settings to control parts of their behavior that are
automated in their templates.

For example the limit on the number of related courses that are picked
up automatically to be displayed on eg. a category page is a tradeoff
between simplicity and performance that depends on the scale of the
Richie instance.

We let Richie users change this through a setting and need it to
be available throughout templates (as it is shared).