✨(courses) add a filter to pick the first N items from a list
The `slice` filter comes stock in Django, and does exactly what we
need to slice queries in templates. It even does more as it uses
the list slicing syntax. However, since that syntax is string based,
it cannot accept variables as part of its argument, without resorting
to verbose hacks in the template.

This is why we decided to add the `first_n` filter, that just implements
the subset of functionality we need, with a simplified signature. This
allows us to pass a variable that defines the length of the slice we
want to take.