💄(layout) implement new integration from fun-corporate work
fun-corporate project developped a new integration more realistic to
real site webdesigns and some improvements about theming.

This is a backport of fun-corporate integration which brings new
templates and Sass sources.

It will obviously break a lots of tests since most are dependant of HTML
structure from templates which have changed a lot. Also, the demo data
will not cover everything that new integration brings, it may be update
further in a new PR.