♻️(courses) move get_es_id from category indexer to model
Using page node path as an ES id is a clever hack that we want to
get away from eventually as it causes issues with Search and
excessive regeneration of ES indices.

Notwithstanding that, it's still used right now. To create links to
the Search page with a search filter activated, we need to a way to
easily grab the ES id from a category, including the "P-" or "L-"
prefix. The easiest way to achieve this is to port the `get_es_id`
method from the categories indexer to the category model.