✨(search) do not delete unrelated indices during regenerate_indices
We just added a setting to make it easy to operate several Richie
instances on the same ElasticSearch by avoiding naming conflicts.

We now just have to make sure we only prune existing unaliased
indices that start with our instance's prefix during
regenerate_indices, which was the last blocking problem to solve
this issue.