🐛(search) make sure the correct org is highlighted on courses in Search
Indexation in ElasticSearch used the order of courses as returned by
the `course.get_organizations()` method to get the main organization
for the course. This worked with a past implementation of this method
but does not work with the current implementation, that does not order
the organizations by placeholder position.

The correct method to use to determine the main organization is
`course.get_main_organization`. We updated the courses indexer
helpers that put and retrieve objects to and from the index to
make sure they get the expected organization as highlighted
organization.