💄(layout) restored login buttons in header
Login button (about login, logout, register) in header was disabled from
new integration because it was an unwanted feature from fun-corporate.
But from Richie it is a required feature so we need to get it back.

This commit enable it again, update its stylesheet to fit to layout and
UserLogin component has been updated to remove unecessary button
classes from Bootstrap.