✨(search) add duration and effort to indexed courses in ElasticSearch
Newly designed versions of course glimpses for Richie's v2 include
duration and effort in the space that was previously dedicated to
the state and call to action. This means we need to index them
along with courses' source to make them available on course search
requests on the API.

We did this in a straightforward way as they are just static (albeit
localized) data on the course, and not included as part of Richie's
course filters for now.