🚸(toolbar) improve plugin description displayed in side toolbar
The name and description of plugins below a placeholder in the side
toolbar was not helpful. We should display the page title preceded by
an indication of the type of the plugin that should be as short as
possible to leave as much space as possible for the page title.
Hence the use abbreviations.