🐛(courses) fix ghost placeholders due to faulty template tags
We thought we could use the same placeholder template tags either on the
current page as editable placeholders or on another page to only display
content from another related page. It worked but created some ghost
placeholders (placeholders from the other page appeared in the
side bar of the current page and were luckily not functional).
We fix it here by using different placeholders for each of these 2 cases.
This is a heavy refactoring but with no impact on existing sites.