♻️(core) remove context processor that was added since latest release
Since the latest release the frontend context processor was added.
It constitutes a breaking change because it is required.
In fact we prefer to group all context variables under one single
existing richie context processor so that projects using richie don't
need to add multiple context processors that are all required.