🐛(frontend) fix an issue with full-text search in course search
Course search had an issue where filters were lost when the user
typed in a full-text search. This was due to the use of a ref in
the <SearchSuggestField /> component which resulted in an outdated
function with an old closure being used.

We had some issues when trying to build tests to show this issue
was resolved. We had to refactor the tests to handle history
mocking differently and stop mocking our own History providing hook,
mocking only the browser's interface.