🔥(section) remove deprecated section template "section_list"
This section template has been removed since it was using a very special
section usage to build the footer menu. Footer menu is now intended to be build
with NestedItem plugin.

This contains a migration to update section using removed template with the
default one, a note about it in UPGRADE document and demo-site command has been
changed to build the footer menu with NestedItem plugin.