🐛(search) fix an issue with whitespace in short search strings
Our full-text search API endpoint only accepts strings of 3 characters
or more. This does not include white space. For instance, "ric" is a
valid search query, but "ri " is not.

This was not reflected in our frontend check, which just looked at the
string length. We just have to trim the string before doing this check
to ensure the query is valid.