👌(section) add new section template "tiles"
In fact a previously removed section template, we finally have no better
way to do this. The purpose is to expose plugin glimpses in a tiny size
with optional summary text.

So finally we add it again but in more flexible way. Currently only
category glimpse is available but more glimpse could be easily added when
required.

This add a new section template and adapted previous tile section CSS.