✨(styleguide) add styleguide page
We need a styleguide to demonstrate base CSS properties and perform
some visual tests.

This just add a simple styleguide system to demonstrate basic common
stylesheet features. It bring a new template in demo, a stylesheet
part and configuration mixin to include in settings which define some
integration properties to display in styleguide.