🐛(section) remove useless dependancy to nesteditem migration
Migration "0005_migrate_sectionlist_to_nesteditem" has been done
planning to migrate section with deprecated template to a nesteditem
which was finally canceled, but there is still a dependancy to
nesteditem in this migration.

Also this dependancy cause trouble on migration on existing database.

This commit just remove the dependancy that should fix the issue.