✨(courses) limit number of related objects on all detail pages
On a real site, some pages may have a large number of related objects.
For example a category may be related to more that 100 courses. In
these cases, the page was too slow to load and too busy for a good
user experience. We use dj-pagination to paginate related objects
thanks to its multipagination feature and with a smart use of html
element ids.