✨(search) improve query analyzer language selection
Each field defines its own default search analyzer in its mapping
(or defaults to the content analyzer if a specific search query
analyzer is not defined). This is appropriate in principle but it
breaks down when we attempt to do multilingual searches: search
queries are then analyzed using the field's analyzer or search
analyzer, which might be in the wrong language, and give terrible
results for the search query.

To prevent this from happening, we decided to enforce an analyzer
at query time. Our approach for now is to lean on the current UI
language to determine the most appropriate analyzer. Simply add it
to the ES query and in our opiniont 99%+ of queries will then use
the most correct analyzer we have.

For edge cases, depending on their number and severity, we'll wait
and see which problems arise in real world use to find the best way
to help our users wrt. their query language when it is not the same
as the current UI language.