♻️(frontend) refactor useCourseSearchParams return value
useCourseSearchParams was returning an array of the params
themselves, and a dispatcher to issue an update for the params.
This is a classic React pattern and it worked well until now.

However, we want to add a third returned value, the previous
value of the params. It's not good design in our opinion to
return more than 2 values of differents types in an array returned
by a hook. We can use an object instead, which will make destructuring
more explicit.