✨(frontend) add last dispatch actions to history state
We are responsing to a new need from more than one part of the app:
knowing what was the last action to run and change the course search
params.

We don't want to duplicate the hack that consists of looking at the
changes and guessing what just happened. Instead, we would rather
just keep the last actions to happen in our state. As we use history
for that (to share accross different component trees), we decided to
expand our history state to include this new information. We nested
it and added the name of the view so we can build upon this expanded
history state in the future.

As this history state is at the core of our frontend app, we had to
updated a lot of frontend tests to make sure everything remains
consistent and works as we intended it.