✨(frontend) scroll course search to top on most user interactions
We previously only scrolled to the top of the course search view
when the user interacted with the pagination. We decided to
experiment with scrolling on any interaction that does not involve
the <SearchSuggestField /> as we hypothesized that most interactions
with filters would be one shots (eg. the user adds one filter, wants
to see results from the top, as opposed to adding multiple filters
at once).