♻️(frontend) change the way SearchSuggestField clears itself
<SearchSuggestField /> needs to clear itself of input text when
another component removes the query from the course search params.
Currently, that only happens with the 'FILTER_RESET' action.

Previously, it kept a ref to the previous version of course search
params and guessed if 'FILTER_RESET' happened to know if it needed
to clear itself.

Now that the useCourseSearchParams hook keeps track of the actions
from the last dispatch, we can expose these from the hook and use
them in <SearchSuggestField /> to know 'FILTER_RESET' was called and
the input needs to be cleared.