♻️(frontend) use an enum for course search params actions
As we worked on changes related to course search params actions,
we realizer their types amounted to magic strings sprinkled all
over the appp.

We decided it would be as good a time as any to refactor them into
members of an enum so they're more clearly associated with their
origin and use.