⚡️(backend) add memcached as a cache backend for Django
By default Django just stores cache in memory. This is not ideal
given the scale (in visitors as well as number of pages) we expect
for Richie instances which serve as a gateway to online education
portals.

The recommended way to handle this is to use memcached instead. We
can add it to our compose stack and add a python + system dependency
to connect with it.