💄(plugins) large_banner hero variant refinement
We need to fit a little bit more closer to the mockups for the hero
variant.

Title field is not used anymore from this variant template. We
expect the title to be add in content and apply a classname
"hero-intro__title" instead. We level up the CKEditor config for
large_banner plugin so now we can add title elements and edit HTML.

Finally, updated demo, added some CSS changes, added
alternative color for strong element in title and pushed content
near to the top and search bar near to the bottom.