👌(layout) fix glitch with "scaleX" usage on wave decoration
There is a bug with CSS transform "scaleX(-1)" (to horizontally flip
an element) when browser view is zoomed. This is visible only on
category and organization details banner, the wave decoration is
displaced for one pixel and so does not cover completely the
background image.

Since this is a bug with from every browser we are not able to patch
it directly. So we just removed the scaleX usage and wave go back to
the original direction as everywhere else.