🏷️(courses) add type definitions for course runs & enrollments
Types are becoming unnecessarily verbose to import throughout the
Richie frontend, especially as many components import several
different types.

This is why we are consolidating these new types for newly available
objects on the API on an index file. We'll be moving other model
based types to this index later on.