✨(frontend) add a component to enroll users in course runs from Richie
A key goal of Richie is to enable integrations with LMSes, and to handle
them as seamlessly as possible from the point of view of the user. What
they would expect to see on our course & course run pages in Richie is
a button that enrolls them in their desired course runs in one click if
possible. Or gives the relevant information otherwise.

It turns out this button needs quite a bit of information to render itself,
and also has many different possible states. To manage this complexity, we
used a state machine to model the whole lifecycle of this button, from
initial load and initial state on page load, to all user interactions.