👌(layout) fix accordion border line
Accordion left border which make the path between item dots was
ending too far compared to mockup. This was due to CSS techniques
and HTML structure.

This commit just fix this with additional line in accordion
Javascript code to add an attribute on item so we can match
something in CSS and avoid border line to continue after the dot
when not opened. Also this adjust some margin on checkmark list from
previous commit 5d3b383.